+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CHANGELOG
   - Updated Play Services and Firebase to 10.0.0.
   - Updated Stetho to 1.4.1.
 - Used Firebase for analytics and crash reporting.
+- Supported Firebase auth (with Google account).
 
 #### v1.15.3 (2016-11-18)
 - Updated deps: